bootloader: Move boot/validation code (boot_from()) into subsys.
Add to a new file (bl_boot) and rename the function to
bl_validate_and_boot().
